add admin

+++ src/main/java/com/example/sweater/repos/MessageRepo.java
@@ -5,11 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-/**
- * Created by HoustoN
- * Date: 9/27/2020
- */
-public interface MessageRepo extends CrudRepository<Message, Integer> {
+public interface MessageRepo extends CrudRepository<Message, Long> {
 
     List<Message> findByTag(String tag);
